requirements.txt
ssh-keygen -t rsa -C "youremail@example.com"adium-theme-ubuntu==0.3.4
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
configparser==3.5.0
cycler==0.10.0
Cython==0.27.3
decorator==4.0.6
Django==1.11.6
dominate==2.3.1
entrypoints==0.2.3
enum==0.4.6
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
html5lib==0.9999999
image==1.5.16
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==7.0.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Markdown==2.2.0
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
mock==2.0.0
nbconvert==5.3.0
nbformat==4.4.0
notebook==5.0.0
numpy==1.13.3
olefile==0.44
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==4.3.0
pkg-resources==0.0.0
progressbar==2.3
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5
pycurl==7.43.0
Pygments==2.2.0
pyliblzma==0.5.3
pyparsing==2.2.0
pypng==0.0.18
pysqlite==1.0.1
Pyste==0.9.10
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
QDarkStyle==2.3.1
qtconsole==4.3.1
rpm-python==4.12.0.1
scandir==1.5
scipy==0.19.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
tensorflow-gpu==1.2.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
tornado==4.5.2
traitlets==4.3.2
unity-lens-photos==1.0
urlgrabber==3.9.1
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
yum-metadata-parser==1.1.4
